model/code/requirements.txt
# models/requirements.txt
torch==2.0.1
torchvision==0.15.2
Pillow==9.5.0